+++ VpnHood.Client.App/AppConnectionState.cs
@@ -1,0 +1,14 @@
+﻿using System.Text.Json.Serialization;
+
+namespace VpnHood.Client.App
+{
+    [JsonConverter(typeof(JsonStringEnumConverter))]
+    public enum AppConnectionState
+    {
+        None,
+        Diagnosing,
+        Connecting,
+        Connected,
+        Disconnecting,
+    }
+}