+++ VpnHood.Client/Exceptions/NoStableVpnException.cs
@@ -4,6 +4,6 @@ namespace VpnHood.Client.Exceptions
 {
     public class NoStableVpnException : Exception
     {
-        public NoStableVpnException() : base("Vpn is connected but it looks the connection is not stable!") { }
+        public NoStableVpnException() : base("Vpn was connected but it looked the connection was not stable!") { }
     }
 }